% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic-misc.R
\name{accuracy}
\alias{accuracy}
\title{Compute predictive accuracy.}
\usage{
accuracy(x, y)
}
\arguments{
\item{x}{A vector of predicted labels.}

\item{y}{A vector of true labels.}
}
\description{
Compute predictive accuracy.
}

