% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-bn-operate.R
\name{AIC.bnc_bn}
\alias{AIC.bnc_bn}
\title{Compute penalized log-likelihood of the Bayesian network classifier.}
\usage{
\method{AIC}{bnc_bn}(object, ...)
}
\arguments{
\item{object}{A \code{bnc_bn} object.}

\item{...}{A data frame.}
}
\description{
The Akaike information criterion (AIC) score:
}
\details{
\eqn{log P(\mathcal{D} \mid \theta) - \frac{1}{2} |\theta|},

where \eqn{|\theta|} is the number of free parameters.
This function must be passed a data frame argument - does not matter how it
is named - in addition to the model (\code{object}). \code{object} must have
its structure and parameters specified.
}

