% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-aode.r
\name{aode}
\alias{aode}
\title{Learn an AODE ensemble.}
\usage{
aode(class_var, features, m = 1, ...)
}
\arguments{
\item{m}{integer The minimum frequency of an value \eqn{x_i} in order to use 
SPODE \eqn{i} in the ensemble.}
}
\value{
\code{bnc_aode} or \code{bnc_str}
}
\description{
If there is a single predictor then returns a naive Bayes.
}
\keyword{internal}
