% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-bn-operate.R
\name{arc_posterior}
\alias{arc_posterior}
\title{Returns the posterior of each arc from the class according to the MANB
method.}
\usage{
arc_posterior(x)
}
\arguments{
\item{x}{The \code{bnc_bn} object.}
}
\description{
The posterior must have already been computed for \code{x}.
}
\references{
Dash D and Cooper GF (2002). Exact model averaging with naive
  Bayesian classifiers. \emph{19th International Conference on Machine
  Learning (ICML-2002)}, 91-98.
}

