% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-assert.R
\name{are_factors}
\alias{are_factors}
\title{Checks if all columns in a data frame are factors.}
\usage{
are_factors(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\description{
Checks if all columns in a data frame are factors.
}
\keyword{internal}
