% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-probs.R
\name{are_pdists}
\alias{are_pdists}
\title{Returns \code{TRUE} is \code{x} is a valid probability distribution.}
\usage{
are_pdists(x)
}
\description{
Returns \code{TRUE} is \code{x} is a valid probability distribution.
}
\keyword{internal}
