% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-dag.R
\name{as_graphNEL}
\alias{as_graphNEL}
\title{To graphNEL.}
\usage{
as_graphNEL(x)
}
\arguments{
\item{x}{a \code{\link{bnc_dag_object}} object. The Bayesian network structure.}
}
\description{
To graphNEL.
}

