% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{augment_arcs}
\alias{augment_arcs}
\title{Returns augmenting arcs.}
\usage{
augment_arcs(bnc_dag)
}
\value{
a character matrix. NULL if no arcs can be added.
}
\description{
Returns augmenting arcs.
}
\keyword{internal}
