% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{augment_bn}
\alias{augment_bn}
\title{Arcs that do not invalidate the tree-like structure}
\usage{
augment_bn(bnc_dag, ...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Arcs that do not invalidate the tree-like structure
}
\keyword{internal}
