% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{augment_kdb}
\alias{augment_kdb}
\title{Arcs that do not invalidate the k-DB structure}
\usage{
augment_kdb(bnc_dag, ...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Arcs that do not invalidate the k-DB structure
}
\keyword{internal}
