% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{augment_kdb_arcs}
\alias{augment_kdb_arcs}
\title{Returns augmenting arcs that do not invalidate the k-DB.}
\usage{
augment_kdb_arcs(bnc_dag, k)
}
\value{
a character matrix. NULL if no arcs can be added.
}
\description{
Returns augmenting arcs that do not invalidate the k-DB.
}
\keyword{internal}
