% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{augment_ode_arcs}
\alias{augment_ode_arcs}
\title{Returns augmenting arcs that do not invalidate the ODE.}
\usage{
augment_ode_arcs(bnc_dag)
}
\value{
a character matrix. NULL if no arcs can be added.
}
\description{
Returns augmenting arcs that do not invalidate the ODE.
}
\keyword{internal}
