% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-params-awnb.R
\name{awnb}
\alias{awnb}
\title{Compute feature weights according to the AWNB method.}
\usage{
awnb(class, dataset, bootstrap_size = 0.5, trees = 10)
}
\description{
Compute feature weights according to the AWNB method.
}
\references{
Mark Hall (2004). A decision tree-based attribute weighting
  filter for naive Bayes. \emph{Knowledge-based Systems}, \bold{20}(2),
  120-126.
}

