% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-params.R
\name{bnc}
\alias{bnc}
\title{Learn a structure and parameters.}
\usage{
bnc(dag_learner, class, dataset, smooth, dag_args = NULL)
}
\arguments{
\item{dag_learner}{A character. Name of the structure learning function.}

\item{class}{A character. Name of the class variable.}

\item{dataset}{The data frame from which to learn network structure and
parameters.}

\item{smooth}{A nonnegative numeric. The smoothing value for Bayesian
parameter estimation.}

\item{dag_args}{A list. Optional additional arguments to \code{dag_learner}.}
}
\description{
A convenience function to learn the structure and parameters in a single
call.
}

