% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bncs.R
\name{bnc_aode}
\alias{bnc_aode}
\title{Returns a \code{c("bnc_aode", "bnc")} object.}
\usage{
bnc_aode(models, class_var, features)
}
\description{
Returns a \code{c("bnc_aode", "bnc")} object.
}
\keyword{internal}
