% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bncs.R
\name{bnc_aode_str}
\alias{bnc_aode_str}
\title{Returns a \code{c("bnc_aode_str", "bnc")} object.}
\usage{
bnc_aode_str(models, class_var, features)
}
\description{
Returns a \code{c("bnc_aode_str", "bnc")} object.
}
\keyword{internal}
