% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-dag.R
\name{bnc_dag}
\alias{bnc_dag}
\title{Creates an augmented naive Bayes with structure but no parameters.}
\usage{
bnc_dag(dag, class)
}
\description{
Creates an augmented naive Bayes with structure but no parameters.
}

