% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bncs.R
\name{bnc_multinet_tan}
\alias{bnc_multinet_tan}
\title{Returns a \code{c("bnc_multinet", "bnc")} object.}
\usage{
bnc_multinet_tan(class, dataset, features, scores)
}
\description{
Returns a \code{c("bnc_multinet", "bnc")} object.
}
\keyword{internal}
