% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-misc.R
\name{bootstrap_ss}
\alias{bootstrap_ss}
\title{Return a bootstrap sub-sample.}
\usage{
bootstrap_ss(dataset, proportion)
}
\arguments{
\item{dataset}{a \code{data.frame}}

\item{proportion}{numeric given as fraction of \code{dataset} size}
}
\description{
Return a bootstrap sub-sample.
}
\keyword{internal}
