% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R
\docType{data}
\name{car}
\alias{car}
\title{Car Evaluation Data Set.}
\format{A \code{data.frame} with 7 columns and 1728 rows.}
\source{
\url{http://goo.gl/GTXrCz}
}
\description{
Car Evaluation Data Set.
}

