% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnc-dag.R
\name{check_dag_class}
\alias{check_dag_class}
\title{Checks the dag and class are adequate for a Bayesian network classifier.}
\usage{
check_dag_class(dag, class)
}
\description{
Checks the dag and class are adequate for a Bayesian network classifier.
}

