% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-input.R
\name{check_dataset}
\alias{check_dataset}
\title{Checks it is a data frame, with all named unique factor columns.}
\usage{
check_dataset(dataset)
}
\description{
Checks it is a data frame, with all named unique factor columns.
}

