% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-mlr.R
\name{check_mlr_attached}
\alias{check_mlr_attached}
\title{Checks if mlr attached.}
\usage{
check_mlr_attached()
}
\description{
mlr must be attached because otherwise  `getMlrOptions()` in `makeLearner` will not be found.
}
\keyword{internal}
