% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-statistics.R
\name{cmi_degrees_freedom}
\alias{cmi_degrees_freedom}
\title{Compute the degrees of freedom for a I(X;Y | Z) test for singleton X, Y, Z.}
\usage{
cmi_degrees_freedom(freqs_table)
}
\description{
Compute the degrees of freedom for a I(X;Y | Z) test for singleton X, Y, Z.
}

