% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-statistics.R
\name{cmi_table}
\alias{cmi_table}
\title{Returns the conditional mutual information three variables.}
\usage{
cmi_table(xyz_freqs, unit = "log")
}
\description{
Returns the conditional mutual information three variables.
}
\keyword{internal}
