% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-graph.r
\name{complete_graph}
\alias{complete_graph}
\title{Returns a complete unweighted graph with the given nodes.}
\usage{
complete_graph(nodes)
}
\arguments{
\item{nodes}{A character vector.}
}
\value{
a \code{graphNEL} object.
}
\description{
Returns a complete unweighted graph with the given nodes.
}
\keyword{internal}
