% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-anb.R
\name{compute_anb_log_joint_per_class}
\alias{compute_anb_log_joint_per_class}
\title{Computes the log joint probability of the observed features for each of the
classes.}
\usage{
compute_anb_log_joint_per_class(x, dataset)
}
\value{
N x nclass numeric matrix .
}
\description{
Assumes that x is an augmented naive Bayes and that data is complete.
}
\details{
If there are no features, dataset should be N x 0 dim.
}
\keyword{internal}
