% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{compute_cll}
\alias{compute_cll}
\title{Computes the conditional log-likelihood of the model on the provided data.}
\usage{
compute_cll(x, dataset)
}
\description{
Computes the conditional log-likelihood of the model on the provided data.
}
\keyword{internal}
