% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap-gRain.R
\name{compute_grain_log_joint}
\alias{compute_grain_log_joint}
\title{Computes the log joint probability of the observed features for each of the classes}
\usage{
compute_grain_log_joint(grain, dataset, class)
}
\description{
Computes the log joint probability of the observed features for each of the classes
}

