% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer.R
\name{compute_ll}
\alias{compute_ll}
\title{Computes log-likelihood of the model on the provided data.}
\usage{
compute_ll(x, dataset)
}
\arguments{
\item{x}{A \code{\link{bnc_bn_object}} object.}

\item{newdata}{A data frame.}
}
\description{
Computes log-likelihood of the model on the provided data.
}

