% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-cpts.R
\name{cpt_1d_values}
\alias{cpt_1d_values}
\title{Returns the name of the first dimensions and the values in the dimension of
the table.}
\usage{
cpt_1d_values(cpt)
}
\description{
Returns the name of the first dimensions and the values in the dimension of
the table.
}

