% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnc-cpts.R
\name{cpt_vars_values}
\alias{cpt_vars_values}
\title{just in their own cpt 1d, not checking in others.}
\usage{
cpt_vars_values(cpts)
}
\description{
just in their own cpt 1d, not checking in others.
}

