% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnc-cpts.R
\name{ctgt2cpt}
\alias{ctgt2cpt}
\title{Turns a contingency table into a conditional probability table}
\usage{
ctgt2cpt(ctgt, smooth)
}
\description{
Turns a contingency table into a conditional probability table
}

