% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Stratified cross validation estimate of predictive accuracy.}
\usage{
cv(x, dataset, k, dag, means = TRUE)
}
\arguments{
\item{x}{List of \code{\link{bnc_bn_object}}. The classifiers to evaluate.}

\item{dataset}{The data frame on which to evaluate the classifiers.}

\item{k}{An integer. The number of folds.}

\item{dag}{A logical. Whether to learn structure on each training set.
Parameters are always learned.}

\item{means}{A logical. Whether to return mean accuracy for each classifier
or to return a k-row matrix with accuracies per fold.}
}
\value{
A numeric vector. The predictive accuracy of each classifier in
  \code{x}. If \code{means = TRUE} then a matrix with k rows.
}
\description{
It may relearn only the parameters on each set of training folds, or relearn
the structure, too. The options used to relearn parameters and, optionally,
structure, are those used to learn the objects in \code{x}.
}

