% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anb-dag.R
\name{dag}
\alias{dag}
\title{Get underlying graph. This should be exported.}
\usage{
dag(x)
}
\arguments{
\item{x}{the bnc object}
}
\description{
Get underlying graph. This should be exported.
}
\keyword{internal}
