% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-statistics.R
\name{extract_ctgt}
\alias{extract_ctgt}
\title{Returns a contingency table over the variables.}
\usage{
extract_ctgt(cols, dataset)
}
\description{
Each variable may be a character vector.
}
\details{
Any rows with incomplete observations of the variables are ignored.
}

