% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-misc.R
\name{fast_equal}
\alias{fast_equal}
\title{Compares all elements in a to b}
\usage{
fast_equal(a, b)
}
\arguments{
\item{b}{numeric. Must be length one but no check is performed.}
}
\description{
Compares all elements in a to b
}
\keyword{internal}
