% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-struct.R
\name{fssj}
\alias{bsej}
\alias{fssj}
\title{Learns Bayesian network classifiers in a wrapper fashion.}
\usage{
fssj(class, dataset, k, epsilon = 0.01, smooth = 0.01)

bsej(class, dataset, k, epsilon = 0.01, smooth = 0.01)
}
\arguments{
\item{class}{A character. Name of the class variable.}

\item{epsilon}{A numeric. Minimum absolute improvement required to keep
searching.}
}
\description{
Learns Bayesian network classifiers in a wrapper fashion.
}
\references{
Pazzani M (1996). Constructive induction of Cartesian product
  attributes. In \emph{Proceedings of the Information, Statistics and
  Induction in Science Conference (ISIS-1996)}, pp. 66-77
}

