% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-params-wanbia.R
\name{get_log_leaf_entries}
\alias{get_log_leaf_entries}
\title{Assuming that the cpt is a leaf, returns 1 instead of a CPT entry when value missing}
\usage{
get_log_leaf_entries(cpt, x)
}
\arguments{
\item{x}{a vector of values}
}
\description{
Assuming that the cpt is a leaf, returns 1 instead of a CPT entry when value missing
}
\keyword{internal}
