% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_connected_components}
\alias{graph_connected_components}
\title{connected_components 
 
 @param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal}
\usage{
graph_connected_components(x)
}
\description{
connected_components 
 
 @param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal
}
