% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_get_adjacent}
\alias{graph_get_adjacent}
\title{Finds adjacent nodes. Has not been tested much}
\usage{
graph_get_adjacent(node, g)
}
\description{
Finds adjacent nodes. Has not been tested much
}
\keyword{internal}
