% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_named_edge_matrix}
\alias{graph_named_edge_matrix}
\title{Returns an edge matrix with node names (instead of node indices).}
\usage{
graph_named_edge_matrix(x)
}
\value{
A character matrix.
}
\description{
Returns an edge matrix with node names (instead of node indices).
}
\keyword{internal}
