% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_subgraph}
\alias{graph_subgraph}
\title{@param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal}
\usage{
graph_subgraph(x)
}
\description{
@param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal
}
