% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-rpart.R
\name{identify_all_testing_depths}
\alias{identify_all_testing_depths}
\title{Identifies all depths at which the features of a classification tree are
tested.}
\usage{
identify_all_testing_depths(tree)
}
\arguments{
\item{tree}{an \code{rpart} object}
}
\value{
a numeric vector. The names are the names of the variables.
}
\description{
Identifies all depths at which the features of a classification tree are
tested.
}
\keyword{internal}
