% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap-rpart.R
\name{learn_unprunned_tree}
\alias{learn_unprunned_tree}
\title{Learns a unprunned \code{rpart} recursive partition.}
\usage{
learn_unprunned_tree(dataset, class)
}
\description{
Learns a unprunned \code{rpart} recursive partition.
}
\keyword{internal}

