% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-chowliu.R
\name{local_ode_score}
\alias{local_ode_score}
\title{Returns pairwise component of ODE (penalized) log-likelihood scores.
In natural logarithms.}
\usage{
local_ode_score(x, y, class, dataset)
}
\description{
Returns pairwise component of ODE (penalized) log-likelihood scores.
In natural logarithms.
}

