% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-chowliu.R
\name{local_ode_score_contrib}
\alias{local_ode_score_contrib}
\title{Returns pairwise component of ODE (penalized) log-likelihood scores. 
In natural logarithms.}
\usage{
local_ode_score_contrib(x, y, class, dataset)
}
\description{
Returns pairwise component of ODE (penalized) log-likelihood scores. 
In natural logarithms.
}
\keyword{internal}
