% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-probs.R
\name{log_normalize}
\alias{log_normalize}
\title{Normalize log probabilities.}
\usage{
log_normalize(lp)
}
\description{
Uses the log-sum-exp trick.
}
\references{
Murphy KP (2012). \emph{Machine learning: a probabilistic
  perspective}. The MIT Press. pp. 86-87.
}
\keyword{internal}
