% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-bn-operate.R
\name{logLik.bnc_bn}
\alias{logLik.bnc_bn}
\title{Compute log-likelihood of the Bayesian network classifier.}
\usage{
\method{logLik}{bnc_bn}(object, ...)
}
\arguments{
\item{object}{A \code{bnc_bn} object.}

\item{...}{A data frame.}
}
\description{
This function must be passed a data frame argument - does not matter how it
is named - in addition to the model (\code{object}). \code{object} must have
its structure and parameters specified.
}
\examples{
data(car)
nb.car <- lp(nb('class', car), car, smooth = 0)
logLik(nb.car, car)
}
\seealso{
\code{\link{logLik}}.
}

