% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-params.R
\name{lpawnb}
\alias{lpawnb}
\title{Learn AWNB parameters.}
\usage{
lpawnb(x, dataset, smooth, trees, bootstrap_size)
}
\arguments{
\item{x}{The \code{\link{bnc_dag}} object. The Bayesian network classifier
  structure.}

\item{dataset}{The data frame from which to learn the classifier.}

\item{smooth}{A numeric. The smoothing value (\eqn{\alpha}) for Bayesian
parameter estimation. Nonnegative.}

\item{trees}{An integer. The number (\eqn{M}) of bootstrap samples to
generate.}

\item{bootstrap_size}{A numeric. The size of the bootstrap subsample,
relative to the size of \code{dataset} (given in [0,1]).}
}
\description{
Deprecated. Use \code{lp} instead.
}

