% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-params-wanbia.R
\name{make_cll}
\alias{make_cll}
\title{Returns a function to compute negative conditional log-likelihood given feature weights}
\usage{
make_cll(class_var, dataset)
}
\description{
Returns a function to compute negative conditional log-likelihood given feature weights
}
\keyword{internal}
