% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-params-wanbia.R
\name{make_cll_gradient}
\alias{make_cll_gradient}
\title{Returns a function to compute the gradient of negative conditional log-likelihood with respect to feature weights}
\usage{
make_cll_gradient(class_var, dataset)
}
\description{
Returns a function to compute the gradient of negative conditional log-likelihood with respect to feature weights
}
\keyword{internal}
