% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{map}
\alias{map}
\title{Assigns instances to the most likely class.}
\usage{
map(pred)
}
\arguments{
\item{pred}{A numeric matrix. Each row corresponds to class posterior 
probabilities for an instance.}

\item{x}{a \code{bnc_fit} object.}
}
\value{
a factor with the same levels as the class variable.
}
\description{
Ties are resolved randomly.
}
\keyword{internal}
