% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/memoise.R
\name{memoise_char}
\alias{memoise_char}
\title{Memoise.}
\usage{
memoise_char(f)
}
\arguments{
\item{f}{a function}
}
\description{
Memoise.
}
\author{
Based on Hadley Wickham's memoise package. memoise function slightly
  modified to avoid the use of digest; the rest functions copied as is from
  memoise.

  Memoise assumes that argument to f is a character vector.
}

