% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R
\name{multinet_tan}
\alias{multinet_tan}
\title{create a multinet using a Chow-Liu's algorithm.}
\arguments{
\item{root}{the  class column of the dataset. The class column will 
divide in diferents class levels  to be used as root of 
the diferent augmenting  trees}
}
\value{
A \code{\link{bnc_dag}} object.
}
\description{
create a multinet using a Chow-Liu's algorithm, by 
maximizing either log-likelihood, the AIC or BIC scores; maximizing 
log-likelihood corresponds to the well-known tree augmented naive Bayes 
(Friedman et al., 1997). When maximizing AIC or BIC the output might be a
forest-augmented rather than a tree-augmented naive Bayes.
}
\examples{
data(car)
ll <- multinet_cl('class', car, , score = 'loglik')   
\dontrun{plot(ll)}
aic <- multinet_cl('class', car, score = 'aic')   
bic <- multinet_cl('class', car, score = 'bic')   
}
\references{
Friedman N, Geiger D and Goldszmidt M (1997). Bayesian network 
  classifiers. \emph{Machine Learning}, \bold{29}, pp. 131--163.
}
