% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-anb.r
\name{named_edge_matrix}
\alias{named_edge_matrix}
\title{Returns an edge matrix with node names (instead of node indices).}
\usage{
named_edge_matrix(g)
}
\arguments{
\item{g}{A \code{\link{graphNEL}}}
}
\value{
A character matrix.
}
\description{
Returns an edge matrix with node names (instead of node indices).
}
\keyword{internal}
