% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anb-families.R
\name{order_acyclic}
\alias{order_acyclic}
\title{Provide an acyclic ordering (i.e., a topological sort).}
\usage{
order_acyclic(families)
}
\description{
Provide an acyclic ordering (i.e., a topological sort).
}
\references{
Beng-Jensen and Gutin, 2007, page 14.
}
\keyword{internal}
