% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-mlr.R
\name{predictLearner.bnc}
\alias{predictLearner.bnc}
\title{predictLearner. Auxiliary mlr function.}
\usage{
predictLearner.bnc(.learner, .model, .newdata, ...)
}
\arguments{
\item{.learner, .model, .newdata}{Internal.}

\item{...}{Internal.}
}
\description{
predictLearner. Auxiliary mlr function.
}
\keyword{internal}
