% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnc-dag-operate.R
\name{print.bnc_dag}
\alias{print.bnc_dag}
\title{Print basic information about a classifier.}
\usage{
\method{print}{bnc_dag}(x, ...)
}
\description{
Print basic information about a classifier.
}
\keyword{internal}
