% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic-graph.R
\name{random_aug_nb_dag}
\alias{random_aug_nb_dag}
\title{Creates a random augmented NB with class as class.}
\usage{
random_aug_nb_dag(class, V, maxpar, wgt)
}
\description{
Creates a random augmented NB with class as class.
}

