% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-assert.R
\name{skip_assert}
\alias{skip_assert}
\title{Whether to do checks or not. Set TRUE to speed up debugging or building.}
\usage{
skip_assert()
}
\description{
Whether to do checks or not. Set TRUE to speed up debugging or building.
}
\keyword{internal}
