% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-aode.r
\name{spode}
\alias{spode}
\title{Returns a Superparent one-dependence estimator.}
\usage{
spode(sp, features, class)
}
\arguments{
\item{sp}{character The superparent.}
}
\description{
Returns a Superparent one-dependence estimator.
}
\keyword{internal}
