% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic-probs.R
\name{sum_log_factors}
\alias{sum_log_factors}
\title{Multiplies a set of factors by converting to log space and then summing.}
\usage{
sum_log_factors(factors)
}
\value{
A numeric matrix. Multiplied in log space.
}
\description{
Multiplies a set of factors by converting to log space and then summing.
}

