% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic-graph.R
\name{superimpose_node}
\alias{superimpose_node}
\title{Adds a node to DAG as root and parent of all nodes.}
\usage{
superimpose_node(dag, node)
}
\description{
Adds a node to DAG as root and parent of all nodes.
}

