% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-hc-operators.R
\name{superparent_children}
\alias{superparent_children}
\title{Return nodes which can be superparents along with their possible children.}
\usage{
superparent_children(bnc_dag)
}
\value{
list of \code{search_state}. NULL if no orphans
}
\description{
Return nodes which can be superparents along with their possible children.
}
\keyword{internal}
