% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R
\docType{data}
\name{voting}
\alias{voting}
\title{Congress Voting Data Set.}
\format{A \code{data.frame} with 17 columns and 435 rows.}
\source{
\url{http://goo.gl/GTXrCz}
}
\description{
Congress Voting Data Set.
}

