% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-anb.R
\name{x2cpt_inds}
\alias{x2cpt_inds}
\title{Maps observations of x to cpt indices. Replicates the indices to cover each
class.}
\usage{
x2cpt_inds(x_cpts, dataset, class, nclass)
}
\value{
A list of integers. The cpt indices of each factor, repeated once for
  each class, and class indices.
}
\description{
Maps observations of x to cpt indices. Replicates the indices to cover each
class.
}
\keyword{internal}
