% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-misc.R
\name{accuracy}
\alias{accuracy}
\title{Compute predictive accuracy.}
\usage{
accuracy(x, y)
}
\arguments{
\item{x}{A vector of predicted labels.}

\item{y}{A vector of true labels.}
}
\description{
Compute predictive accuracy.
}
\examples{
 
data(car)
nb <- bnc('nb', 'class', car, smooth = 1)
p <- predict(nb, car)
accuracy(p, car$class)
}
