% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-struct.R
\name{aode}
\alias{aode}
\title{Learn an AODE ensemble.}
\usage{
aode(class, dataset, features = NULL)
}
\arguments{
\item{class}{A character. Name of the class variable.}

\item{dataset}{The data frame from which to learn the classifier.}

\item{features}{A character vector. The names of the features. This argument
is ignored if \code{dataset} is provided.}
}
\value{
A \code{bnc_aode} or a \code{bnc_dag} (if returning a naive Bayes)
}
\description{
If there is a single predictor then returns a naive Bayes.
}
