% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap-mlr.R
\name{as_mlr}
\alias{as_mlr}
\alias{makeRLearner.bnc}
\alias{predictLearner.bnc}
\alias{trainLearner.bnc}
\title{Bridge to \code{mlr}}
\usage{
makeRLearner.bnc()

as_mlr(x, dag, id = "1")

trainLearner.bnc(.learner, .task, .subset, .weights, ...)

predictLearner.bnc(.learner, .model, .newdata, ...)
}
\arguments{
\item{x}{A \code{\link{bnc_bn_object}} object.}

\item{dag}{A logical. Whether to repeat structure learning on each fold or
just parameter learning.}

\item{id}{A character.}

\item{.learner,.task,.subset,.weights,.model,.newdata}{Internal.}

\item{...}{Internal.}
}
\description{
Only \code{as_mlr} should be called by the user.
}

