% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrap-mlr.R
\name{as_mlr}
\alias{as_mlr}
\title{Convert to \code{mlr}.}
\usage{
as_mlr(x, dag, id = "1")
}
\arguments{
\item{x}{A \code{\link{bnc_bn}} object.}

\item{dag}{A logical. Whether to learn structure on each training subsample.
Parameters are always learned.}

\item{id}{A character.}
}
\description{
Convert a \code{\link{bnc_bn}} to a \code{\link[mlr]{Learner}}
object.
}
\examples{
data(car)
nb <- bnc('nb', 'class', car, smooth = 1)
nb_mlr <- as_mlr(nb, dag = FALSE, id = "ode_cl_aic")
nb_mlr
}

