% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaImplement.R, R/anb-clg-params.R
\name{BetaImplement}
\alias{BetaImplement}
\title{Beta_implement package}
\usage{
BetaImplement(x, dataset)
}
\arguments{
\item{x}{The bnc_dag object. The Bayesian network classifier structure.}

\item{dataset}{The data frame.The data frame from which to learn coefficients.}
}
\value{
bnc_fit_clg. bnc_fit_clg is the object that contains the coefficients of each node in the structure
}
\description{
Learn the coefficients of all the nodes of a structure with the function lm(). Returns a bnc_fit_clg.

More ditailed description
}
\details{
betaImplement learns the coefficient of each node in the bayesian network structure by using the function lm().

BetaImplemet only returns the coefficient for numerical node. In case of categorical node, it will return a error information.

When the parents of the node are categorical + numerical, it will returns the coefficients by using the numerical parents based on the different combination of categorical parents.
For example:

We have a node speed, the parents are velocity(numeric), wheel(categorical,4 levels) and accelerator(categorical, 2 level).
 The node speed will have 8 different result of lm(speed~velocity,data).
Otherwise, it will get another error information.
}
\examples{
structure<-tan_cl('Species',as.data.frame(lapply(iris,as.factor)))
plot(structure)
x<-BetaImplement(structure,iris)
x$Sepal.Width
}
\author{
shanshancheng
}
