% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R, R/anb-bn-operate.R, R/anb-bn.R
\name{bnc_bn_object}
\alias{bnc_bn_object}
\alias{classes}
\alias{nparams}
\alias{params}
\alias{values}
\title{A Bayesian network classifier with structure and parameters.}
\usage{
nparams(x)

params(x)

values(x)

classes(x)
}
\arguments{
\item{x}{The \code{bnc_bn} object.}
}
\description{
You can query this object using the following functions, in addition to all the functions that can be applied to a \code{\link{bnc_dag_object}}.
}
\section{Functions}{
\itemize{
\item \code{nparams}: Returns the number of free parameters in the model.

\item \code{params}: Returns the list of CPTs, in the same order as \code{\link{vars}}.

\item \code{values}: Returns the possible values of each variable, in the same order as \code{\link{vars}}.

\item \code{classes}: Returns the possible values of the class variable.
}}

