% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anb-bn.R, R/bnclassify-doc.R
\name{params}
\alias{bnc_bn_object}
\alias{classes}
\alias{params}
\alias{values}
\title{A Bayesian network classifier with structure and parameters.}
\usage{
params(x)

values(x)

classes(x)
}
\arguments{
\item{x}{The bnc_bn_object}
}
\description{
You can query this object using the following functions, in addition to all the functions that can be applied to a \code{\link{bnc_dag_object}}.
}
\section{Functions}{
\itemize{
\item \code{params}: Returns the list of CPTs, in the same order as \code{\link{vars}}.

\item \code{values}: Returns the possible values of each variable, in the same order as \code{\link{vars}}.

\item \code{classes}: Returns the possible values of the class variable.
}}

