% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnclassify-doc.R
\docType{package}
\name{bnclassify}
\alias{bnclassify}
\alias{bnclassify-package}
\title{Algorithms for learning Bayesian network classifiers from data.}
\description{
Algorithms for learning Bayesian network classifiers from data.
}
\references{
Bielza C and Larra\~{n}aga P (2014), Discrete Bayesian network
  classifiers: A survey. \emph{ACM Computing Surveys}, \bold{47}(1), Article
  5.

  Friedman N, Geiger D and Goldszmidt M (1997). Bayesian network classifiers.
  \emph{Machine Learning}, \bold{29}, pp. 131--163.
}

