% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R
\docType{data}
\name{car}
\alias{car}
\title{Car Evaluation Data Set.}
\format{A \code{data.frame} with 7 columns and 1728 rows.}
\source{
\url{http://goo.gl/GTXrCz}
}
\description{
Data set from the UCI repository:
\url{https://archive.ics.uci.edu/ml/datasets/Car+Evaluation}.
}
