% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-statistics.R
\name{cmi}
\alias{cmi}
\title{Returns the (conditional) mutual information among (three) two sets of variables.}
\usage{
cmi(x, y, dataset, z = NULL, unit = "log")
}
\description{
Returns the (conditional) mutual information among (three) two sets of variables.
}

