% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{compute_ll}
\alias{compute_ll}
\title{Computes log-likelihood of the model on the provided data.}
\usage{
compute_ll(x, dataset)
}
\arguments{
\item{x}{A \code{\link{bnc_bn}} object.}

\item{dataset}{A data frame.}
}
\description{
Computes log-likelihood of the model on the provided data.
}
\keyword{internal}
