% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anb-cpts.R
\name{cpt_vars_values}
\alias{cpt_vars_values}
\title{Get just form first dimension in their own cpt, not checking for consistency
in others.}
\usage{
cpt_vars_values(cpts)
}
\description{
Get just form first dimension in their own cpt, not checking for consistency
in others.
}
\keyword{internal}
