% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{cv}
\alias{cv}
\title{Stratified cross validation estimate of predictive accuracy.}
\usage{
cv(x, dataset, k, dag, smooth = NULL)
}
\arguments{
\item{x}{List of \code{\link{bnc_bn_object}}. The classifiers to evaluate.}

\item{dataset}{The data frame on which to evaluate the classifiers.}

\item{k}{An integer. The number of folds.}

\item{dag}{A logical. Whether to learn structure on each training set.
Parameters are always learned.}

\item{smooth}{A numeric. The smoothing value (\eqn{\alpha}) for Bayesian
parameter estimation. Nonnegative.}
}
\value{
A numeric vector. The predictive accuracy of each classifier in
  \code{x}.
}
\description{
Stratified cross validation estimate of predictive accuracy.
}

