% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-anb.r
\name{direct_forest}
\alias{direct_forest}
\title{Direct an undirected graph.}
\usage{
direct_forest(g, root = NULL)
}
\arguments{
\item{g}{An undirected graph.}

\item{root}{A character. Optional tree root.}
}
\value{
A directed graph
}
\description{
Starting from a \code{root} not, directs all arcs away from it and applies 
the same, recursively to its children and descendants. Produces a directed
forest.
}
\keyword{internal}
