% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-anb.r
\name{direct_tree}
\alias{direct_tree}
\title{Direct an undirected graph.}
\usage{
direct_tree(g, root = NULL)
}
\value{
A graph. The directed tree.
}
\description{
The graph must be connected and the function produces a directed tree.
}
\keyword{internal}
