% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_add_edges}
\alias{graph_add_edges}
\title{Add edges
Does not allow edges among adjacent nodes}
\usage{
graph_add_edges(from, to, g)
}
\description{
Add edges
Does not allow edges among adjacent nodes
}
\keyword{internal}
