% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_subgraph}
\alias{graph_subgraph}
\title{Subgraph.  
 Only for a directed graph?
 @param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal}
\usage{
graph_subgraph(nodes, x)
}
\description{
Subgraph.  
 Only for a directed graph?
 @param  x currently a graphNEL. TODO But will be a graph_internal.
 @keywords internal
}
