% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-internal.r
\name{graph_subgraph}
\alias{graph_subgraph}
\title{Subgraph.  
Only for a directed graph?}
\usage{
graph_subgraph(nodes, g)
}
\arguments{
\item{nodes}{character}

\item{x}{currently a graphNEL. TODO But will be a graph_internal.}
}
\description{
Subgraph.  
Only for a directed graph?
}
\keyword{internal}
