% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontend-anb.r
\name{graph_union}
\alias{graph_union}
\title{Merges multiple disjoint graphs into a single one.}
\usage{
graph_union(g)
}
\arguments{
\item{g}{A \code{\link{graphNEL}}}
}
\value{
A \code{\link{graphNEL}}.
}
\description{
Merges multiple disjoint graphs into a single one.
}
\keyword{internal}
