% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-rpart.R
\name{identify_min_testing_depths}
\alias{identify_min_testing_depths}
\title{Identifies the lowest (closest to root) depths at which the features of a 
classification tree are tested.}
\usage{
identify_min_testing_depths(tree)
}
\description{
Identifies the lowest (closest to root) depths at which the features of a 
classification tree are tested.
}
\keyword{internal}
