% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-rpart.R
\name{learn_unprunned_tree}
\alias{learn_unprunned_tree}
\title{Learns a unpruned \code{rpart} recursive partition.}
\usage{
learn_unprunned_tree(dataset, class)
}
\description{
Learns a unpruned \code{rpart} recursive partition.
}
\keyword{internal}
