% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{memoise_char}
\alias{memoise_char}
\title{Memoise a function.}
\usage{
memoise_char(f)
}
\arguments{
\item{f}{a function}
}
\description{
Based on Hadley Wickham's memoise package. Assumes that argument to f is a
character vector.
}
\details{
This function is a slightly modified version of
\code{memoise} to avoid the use of digest. The rest functions
copied as is from memoise.
}
\author{
Hadley Wickham, Bojan Mihaljevic
}
\keyword{internal}
