% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0bnclassify-doc.R, R/learn-struct.R
\name{multinet_cl}
\alias{multinet_cl}
\title{Create an ensemble of Bayesian network using a Chow-Liu's algorithm.}
\usage{
multinet_cl(class, dataset, features = NULL, score = "loglik")
}
\arguments{
\item{class}{A character. Name of the class variable.}

\item{dataset}{The data frame from which to learn the classifier.}

\item{features}{A character vector. The names of the features. This argument
is ignored if \code{dataset} is provided.}

\item{root}{the  class column of the dataset. The class column will 
divide in diferents class levels  to be used as root of 
the diferent augmenting  trees}
}
\value{
A \code{\link{bnc_dag}} object.

A \code{multinet_cl} or a \code{bnc_dag} (if returning a naive Bayes)
}
\description{
create an ensemble of Bayesian network  using a Chow-Liu's algorithm, by 
maximizing either log-likelihood, the AIC or BIC scores; maximizing 
log-likelihood corresponds to the well-known tree augmented naive Bayes 
(Friedman et al., 1997).

If there is a single predictor then returns a naive Bayes.
}
\examples{
data(car)
ll <- multinet_cl('class', car)
ll <- multinet_cl('class', car, score = 'loglik')      
\dontrun{plot(ll)}
aic <- multinet_cl('class', car, score = 'aic')   
bic <- multinet_cl('class', car, score = 'bic')   
}
\references{
Friedman N, Geiger D and Goldszmidt M (1997). Bayesian network 
  classifiers. \emph{Machine Learning}, \bold{29}, pp. 131--163.
}
