% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-graph.R
\name{named_edge_matrix}
\alias{named_edge_matrix}
\title{Returns an edge matrix with node names (instead of node indices).}
\usage{
named_edge_matrix(g)
}
\arguments{
\item{g}{A graphNEL.}
}
\value{
A character matrix.
}
\description{
TODO Not sure this function is used.
}
\keyword{internal}
