% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn-struct.R
\name{nb}
\alias{nb}
\title{Learn a naive Bayes network structure.}
\usage{
nb(class, dataset = NULL, features = NULL)
}
\arguments{
\item{class}{A character. Name of the class variable.}

\item{dataset}{The data frame from which to learn the classifier.}

\item{features}{A character vector. The names of the features. This argument
is ignored if \code{dataset} is provided.}
}
\description{
Learn a naive Bayes network structure.
}
\examples{
data(car)
nb <- nb('class', car)
nb2 <- nb('class', features = letters[1:10])
plot(nb2)
}

