% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnc-dag-operate.R
\name{plot.bnc_dag}
\alias{plot.bnc_dag}
\title{Plot the structure.}
\usage{
plot.bnc_dag(x, y, layoutType = "dot", ...)
}
\arguments{
\item{x}{The bnc_dag_object}

\item{y}{Not used}

\item{layoutType}{Layout type}

\item{...}{Not used.}
}
\description{
Plot the structure.
}

