% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-mlr.R
\name{trainLearner.bnc}
\alias{trainLearner.bnc}
\title{trainLearner. Auxiliary mlr function.}
\usage{
trainLearner.bnc(.learner, .task, .subset, .weights, ...)
}
\arguments{
\item{.learner, .task, .subset, .weights}{Internal.}

\item{...}{Internal.}
}
\description{
trainLearner. Auxiliary mlr function.
}
\keyword{internal}
