% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bnclassify-doc.R
\docType{data}
\name{voting}
\alias{voting}
\title{Congress Voting Data Set.}
\format{A \code{data.frame} with 17 columns and 435 rows.}
\source{
\url{http://sourceforge.net/projects/weka/files/datasets/UCI and StatLib/uci-20070111.tar.gz}
}
\description{
Congress Voting Data Set.
}

